android vulkan: disallow noflags memory types again for AHB (for now)
bug: 134773396

This CL makes BasicVulkanGpuTests pass again, via disallowing NVIDIA's
sysmem memory types (as they dont work well with dedicated/external).

The better (upcoming) solution is to query in the beginning what memory
types correspond to dedicated + external on the host in the first place,
instead of guessing based on heuristics.

Change-Id: Ic259eecd4cb8c2e38a5737c98ac22837febf7896