vkGetPhysicalDeviceProperties: Set device type to Virtual GPU.
Vulkan has a specific type VK_PHYSICAL_DEVICE_TYPE_VIRTUAL_GPU
to indicate if a device is emulated or a virtual node.
We can overwrite the physical device type of all emulated
Vulkan physical devices so that clients will be aware that
they are running in emulator.

vkgen scripts updated in
https://android-review.googlesource.com/c/platform/external/qemu/+/1350142

Bug: fxb/54086
Change-Id: I402f9b0f9266eb5e55d744a04294a0d971e7ed80