codecs: enable vpx parallel decoding
We are doing decoding in parallel on host, so
turn on the flag in the guest so it can flush
to get the remaining frames out of host hardware
decoder. Otherwise, we might lose 1 or more frames
in the end, especially cuda decoder.

BUG: 160616123

Test:
with gpu decode on
atest android.media.cts.EncodeDecodeTest#testVP8EncodeDecodeVideoFromSurfaceToSurfaceQCIF
atest com.google.android.exoplayer.gts.DashTest#testWidevineVp9*

Change-Id: If7ec8131d1fd13f478ed197c9e139d4d31b16ff2