Always inc/decref in sync(CommandBuffers/Queues)
Also remove the decref in unregister cmdbuf/queue.

Also make it not a pointer.

Also don't inc/dec the m_stream member along with vkencoder inc/decrefs
there is supposed to be only one m_stream inc when encoder is created
and then a dec ref when encoder is deleted...

Change-Id: I3948cf94ee7d399725b0ef1edd01cdf9d78fe777