fuchsia: Implement vkGetBufferCollectionProperties2FUCHSIA
This change implements vkGetBufferCollectionProperties2FUCHSIA.
This adds more fields to the old VkBufferCollectionProperties,
including the index of createInfo used to allocate memory,
formatFeature flags, color space info and YUV-images related
info.

Before the properties is returned to clients, it will be
duplicated as a shallow copy in info_VkBufferCollectionFUCHSIA
for other Vulkan functions to use.

Change-Id: I8e726421f2b03ee15cf471a116dbb7386f79c44f