fuchsia: Set tiling to linear if sysmem image uses host-visible memory.
For sysmem-backed images, if it uses host-visible memory,
the VkImage created on host always use linear tiling.
When we create VkImage using the VMO, we should set its
tiling to VK_IMAGE_TILING_LINEAR so that it could be
correctly sampled using linear tiling, while clients don't
need to know the exact tiling of that image.

This fixes sysmem images sampling on Intel and NVIDIA
host GPUs.

Change-Id: I62377795b5add05b577786d146b779365cb58da3