fuchsia: Do not return error if client tries importing non-colorBuffer memory.
This matches the behavior with other Vulkan device drivers on
Fuchsia, where importing a read-only host memory VMO into
Vulkan driver doesn't work, while it is not considered as an
error if trying to do so. Vulkan driver could return a
VkMemoryZirconHandlePropertiesFUCHSIA with memoryTypeBits of 0
indicating that the DeviceMemory cannot be allocated.

TEST=gfx_unittests on FEMU

Change-Id: I38efd07cbc50822517bccf1190cfabee4ee0d963