c2-codecs: use real pts instead of work index
for h264, video tool box on osx does not reorder
the decoder output frames, and a work around
is to buffers the output frames and then feed
to guest using the pts order.

this cl changes the work index to real pts when
calling host to decode each input frame.
and then convert it back to work index for
further processing.

Change-Id: I19b2d6519ba753475b32a48297769f182323f4cb