vk: SyncThread wait for vkQueueSubmit before waiting for fences
Vulkan specs require fences of vkQueueSubmit to be *externally
synchronized*, i.e. we cannot submit a queue while waiting for the
fence in another thread. Thus for a SyncThread executing SYNC_THREAD_
WAIT_VK commands, it has to first wait until a vkQueueSubmit() using
this fence is called before calling vkWaitForFences().

For each VkFence, we'll use a conditional variable and mutex for
thread synchronization.

Also see:
https://www.khronos.org/registry/vulkan/specs/1.2/html/vkspec.html#fundamentals-threadingbehavior
https://github.com/KhronosGroup/Vulkan-LoaderAndValidationLayers/issues/519

Change-Id: I8b7bdbb3dfd9d703a6f962aee9d6ede8fc6dc016