base: Add "fit" library
libfit is a header-only library used in Fuchsia SDK providing
immproved control flow and memory management beyond the C++17
standard library. This includes:

- Move-only function objects: fit::Function / Callback / Closure;
- Move-only RAII deferred action wrapper: fit::defer;
- Thread safety checker and thread-safety annotations.

The library is located under android::base::fit namespace, and is
modified to fit the coding style of android::base libraries.

No-Typo-Check: keep the identifiers the same as those used in the original Fuchsia repository

Change-Id: I785f4c0d22a9b2597097719432dbb0d147bd80d8