[fuchsia] Use zxio_create() to test if services are accessible
The VulkanDevice::IsAccessible routine tests if a connecting to a given
name results in a responsive object. This updates this routine to use
the public zxio_create() entry point which internally tests that the
endpoint is responsive.  The advantage of this approach is it removes
the dependency on the internal zxio/inception.h header in favor of the
public zxio.h entry points.

Change-Id: I43e958b361121e27496aa1440611b573036a9aa2