hwcomposer.ranchu: Maintain its own copy of drmhwcgralloc.h
The ranchu hwcomposer logic pulls in the drmhwcgralloc.h from
drm_hwcomposer for the hwc_drm_bo definition.

However, upstream in drm_hwcomposer, that structure has been
reworked, which then causes the ranchu hwcomposer to stop
building when we update the drm_hwcomposer tree.

As the structure is just being re-utilized as convenience /
storage structure and isn't used as a shared data structure with
other components, so I've copied it over locally so things can
continue to build and the drm_hwcomposer can be updated
independently.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I86e100c46ffe211c9e8c48c0c9222592d41f240a