deepcopy: Always cast pointer to "void*" for allocated pNext structs.
Previously for pNext pointers, we cast allocated pointer to the type
defined in vk.xml. However recent Vulkan header changes some of the
pNext pointers from "const void*" to "void*". This could cause an
incompatibility issue when using goldfish-vulkan with a different
Vulkan header version (which is the case for Fuchsia).

So in order to solve this issue, we always use "void*" for allocated
pNext structs, instead of "const void*".

Change-Id: I70586d92ef6482fbb50df486e69c3f6aeabbfd8d