Add display finder DRM flow
... so that downstream devices of Cuttlefish that run without
Cuttlefish's config server can still initialize displays.

Eventually, we want to enable this on Cuttlefish as well once
Crosvm fully supports the display properties we use (it looks like
it doesn't support DPI and refresh rate yet).

This is kind of a cherry-pick of aosp/2203917 for hwc2 for use in
older branches which don't have hwc3.

Bug: b/239855014
Test: build
Change-Id: I09aed1beb323d540095103f8c48f44f8c9a9a84e