Load DPI from EDIDs preferred detailed timing descriptor
Prior to this change, display physical width and height were
taken from ` drmModeConnector`'s `mmWidth` and `mmHeight`
members. However, these properties ultimately come from
bytes 21 and 22 of the EDID blob which are a single byte
each and represent screen size in centimeters with a range
of 1-255cm. Cuttlefish (phone)'s default size is 720x1280
with a DPI of 320 which gives us a

  (720/320) = 2.25in / 5.715cm / 57.15mm screen.

Packing this into the single EDID byte gives us a 5cm or 6cm
screen and using this to recompute the DPI in the guest gives

  (720px / 5cm) * 2.54 cm-per-inch = 365.76 DPI
  (720px / 6cm) * 2.54 cm-per-inch = 304.8 DPI

Instead, let's take the screen size from the preferred detailed
timing descriptor (the first one)'s bytes 12-14 which give us 12
bits and represent the 'active' video image size in millimeters
with a range of 0-4095mm. Let's assume the preferred detailed
timing descriptor best matches and fully utilizes the entire
screen. Then,

  (720px / 57mm) * 25.4 mm-per-inch = 320.8 DPI

which reconstructs the configured DPI on the host.

Bug: b/163864461
Test: cvd start --display0=width=500,height=500,dpi=<various>
Change-Id: I1503f3a300c3932eab961721dd184467a6ce0bf1