Avoid returning unboxed VkDeviceMemory to the guest when map fails
If mapping of the device memory fails, this was previously still
returning the unboxed host VkDeviceMemory to the guest. In the
guest, vkAllocateMemory() returns VK_ERROR_OUT_OF_HOST_MEMORY
but then the pMemory output would still be populated which leads
to RAII objects (in ANGLE for example) trying to vkFreeMemory()
the unboxed host VkDeviceMemory. This leads to crashes when the
VK decoder trys to unbox the VkDeviceMemory passed to VkFreeMemory().

Bug: b/233939967
Test: cvd start --gpu_mode=gfxstream --enable_gpu_angle
Test: cts -m CtsNativeHardwareTestCases
Change-Id: I1ea4277d4cce3be950df79a34c7392be00f5b43b