(Reland) Route gl<->vk ColorBuffer syncs through FrameBuffer
RenderControl functions such as rcBindTexture(),
rcFlushWindowColorBuffer(), etc act as a kind of boundary between
ColorBuffer usage across APIs (GL/EGL, Vulkan, and Gralloc) and need
to sync the state of the ColorBuffer between the various host APIs
that back ColorBuffers. For example, rcBindTexture() needs to sync
the state of GL texture backing ColorBuffers for GL usage and does
so by reading the state from the VK image.

Upcoming ColorBuffer refactoring, to enable running without host GL,
needs to make the ColorBuffer VK setup path not depend on the
ColorBuffer GL path. With this, setup of the VK backing will no
longer be deferred until the first VK usage and ColorBuffer::create()
is updated to perform both ColorBufferGl::create() and
ColorBufferVk::create() immediately.

However, this can lead to a potential deadlock because of the order
of locks acquired during the ColorBuffer creation flow and the
RenderControl sync'ing flows.

* The RenderControl functions use goldfish_vk::readColorBufferToGl()
  and goldfish_vk::updateColorBufferFromGl() to perform downloads and
  uploads for both GL and VK. These calls lock sVkEmulationLock first
  and then lock FrameBuffer::m_lock second.

* The virtio-gpu resource_create_3d() path calls into
  FrameBuffer::createColorBuffer() which then calls into
  ColorBufferGl::create() and ColorBufferVk::create(). These calls lock
  FrameBuffer::m_lock first and then sVkEmulationLock second.

This change updates the RenderControl functions to go through the
FrameBuffer lock first so that it is kind-of always a top-down locking
approach.

Reland needed to include aosp/2385074.

Bug: b/233939967
Test: android build
Test: cmake build
Test: cvd start --gpu_mode=gfxstream
Test: gfxstream unit tests
Change-Id: Ic4310d89127d489430d64a2b1e3fcc59ba01fab1