vulkan-cereal: reduce usage of emulated format feature bit
It's unlikely the emulate ASTC/ETC2 feature bits will ever
end up as flags to the external stream_renderer interface.
Ideally, formats would be emulated as needed and that's what
the current logic does.  This removes an extra layer of flags
for that.  Developers may simply modify Framebuffer.cpp +
recompile if they wish to disable emulation.

BUG=254065847
TEST=compile

Change-Id: I7057dd553fda8b41ca3209f893cc636179ea203b