(Reland) Use dedicated memory allocations when running with Mesa
... as otherwise Mesa expects linear tiling.

The attached guest side change updates the guest to always pass
through the VkMemoryDedicatedAllocateInfo as the guest does not
yet know if the host is using dedicated allocs for
Buffer/ColorBuffer. Then, VkDecoderGlobalState is updated to
check if Buffer/ColorBuffer used a dedicate alloc and use the
VkMemoryDedicatedAllocateInfo from the guest if needed.

Removes the extra page size alignment/padding of memory in
allocExternalMemory() to avoid VVL errors with dedicated allocs:

  VUID-VkMemoryDedicatedAllocateInfo-image-01433 ...  The Vulkan
  spec states: If image is not VK_NULL_HANDLE,
  VkMemoryAllocateInfo::allocationSize must equal the
  VkMemoryRequirements::size of the image

Reland note: no changes from previous aosp/2385074.

Bug: b/265186355
Test: cvd start --gpu_mode=gfxstream (with AMD GPU)
Test: cvd start --gpu_mode=gfxstream (with Nvidia GPU)
Change-Id: Ie632ae56f98ea20e2ea9a170ea1fec2384067f48