Move assumptions about "main" backing into ColorBuffer
Updates parts of RenderControl and VK emulation which have explicit
GL->VK and VK->GL syncs (and implicitly assumes a default ColorBuffer
"main" backing) to use flushFrom<Api>()/invalidateFor<Api>() so that
the syncs can be hidden  and handled inside of the generic ColorBuffer.

Removes the ReplaceContentsTest test. replaceContents() was really for
"update GL from these bytes from VK". However, the test never actually
guarenteed that the "readColorBuffer()" used to check the ColorBuffer
contents was reading from the GL side of things.

Bug: b/270459313
Test: cvd start --gpu_mode=gfxstream
Test: cvd start --gpu_mode=gfxstream --enable_gpu_angle=true
Test: cts -m CtsGraphicsTestCases
Test: cts -m CtsNativeHardwareTestCases
Test: Open Camera
Test: Open Youtube
Change-Id: If1e1cf9bc0600eebaddeba991fd9e86beab1d9f2