Look up memory properties after updating memory type index
When importing a Buffer/ColorBuffer, the memory type index used by
the guest might be different than the one used by the host. This
is an issue that still needs to be addressed but for now avoid
trying to map something that the guest thinks is host visible but
that the host knows is not host visible.

Bug: b/272079201
Test: cvd start --gpu_mode=gfxstream_guest_angle
Test: cts -m CtsNativeHardwareTestCases
Change-Id: Ie1b94eaf7ecc2cf9b493bdc07dce7ab7f8af7ae4