vulkan-cereal: deferred resource creation host-side implementation
Both the old method (on_vkGetMemoryHostAddressInfoGOOGLE) and
vkGetBlob are routed to vkGetBlobInternal.  The difference is the
old method generated a host ID, returned to the guest.   The new
method assigns a guest ID to host memory during VkAllocateMemory.
When it is time to map the buffer, vkGetBlob(..) can be called.

This avoids mapping host visible memory all the time.  For discrete
GPUs that may not be a huge difference, but for integrated GPUs
we don't want to create blob resources for every host side
allocation.

BUG=262595016
TEST=compile and run CTS

Change-Id: I9b8caff0a603302b95ddd0ab1fab907e8ff70d59