vulkan-cereal: guest memory prototype
This adds initial support for the F_CREATE_GUEST_HANDLE flow. crosvm
creates the handle from guest memory using udmabuf, and gives it
to gfxstream via stream_renderer_create_blob(..).  The guest also
sends a placeholder command buffer and more importantly fence to guage
if the resource creation is complete.  Right now, gfxstream doesn't
signal the fence, but rutabaga does since it was easier.

We may get rid of the placeholder fence in the final version.

After the fence as been created, the handle is imported into the
Vulkan ICD using dmabuf and external memory extensions.

BUG=262595016
TEST=compile and run CTS

Change-Id: I36e4aab9d6af2f0a6f62c8f71ba70d2498e13012