Fix egl include path.
'FetchContent_MakeAvailable' has to be called before using 'egl_SOURCE_DIR' variable.
The issue remained unnoticed because EGL headers can be found on the system.

Also removed the valiable that lists the projects to be fetched.
There is no evidence that is is advantageous to get them all at once.

Change-Id: Ib613c83ad676edf37c699d0b7a944620422386b5
Bug: b/276425028