vulkan-cereal: nuke unnecessary locking
The crosvm virtio-gpu device is single threaded, so is QEMU.
AEMU doesn't use this interface yet, preferring RenderLib.

gfxstream manages the RenderThreads when creating a new context
or guest thread.  This locking is unnecessary.

Though, there is a slight race between the HostmemIdMapping
(accessible by RenderThreads) and the virtio-gpu thread.  We've
been getting lucky there.

BUG=239639751
TEST=compile

Change-Id: Ib60bd822949ee02c2d95a58fbae8e3d6e6548a10