vulkan-cereal: remove unnecessary checks
The comment is somewhat misleading --`transformImpl_VkExternalMemoryProperties_tohost` does no transformations at.

The guest side is careful to append the right struct anyways, so currently `VkExportMemoryAllocateInfo` is appended.

And if the guest did (for example, requesting a Fuchia handle on a
Linux guest), probably the VK driver or validation layers would almost
immediately complain.

Long-term, we could even reduce or even eliminate the goldfish_vk_transform_guest.cpp
and it's 20K LOC, since it deals with memory and external memory.


BUG=202552093
TEST=compile

Change-Id: I1f320b7de2b303a1165a978180d8563a0c1c63ea