Use context specific ring for EXPORT_SYNC_VK
... as the guest assumes fence context support in
ResourceTracker::createFence(). Similar to aosp/2152133, without
this change, the task for the vkWaitForFences() and the fence from
the EXPORT_SYNC_VK virtio gpu execbuffer command end up on
different rings which leads to the fence being signalled almost
immediately. This can lead to the guest trying to re-use resources
(in the case of b/281735034, a VkCommandBuffer) before prior use
has finished.

Bug: b/281735034
Test: cvd start --gpu_mode=gfxstream_guest_angle_host_swiftshader
Change-Id: Iaf16f0d74d9648cb986d8fbe5c7e5ff4e1951809