Bind to gl context only once in post worker
Post thread is supposed to have a dedicated context, we should
bind to it once and never release until the emulator exits.

Note that this does not apply to Mac because Mac post worker runs
on UI thread.

We are hypothesizing that this might be related to an emuator
crash that happens on context rebinding.

Bug: 281872037
Change-Id: I554613c704743be496be5461e2b6ad4403fa4b1a