gfxstream-guest: add precise virtio-gpu capabilities
We need separate virtio-gpu capabilities for this features, for
separate versioning.  As of right now, versioning is done via
rcGetHostExtensionsString, which calls rcGetGLString host side.

We should probably aim for separate versioning for GLES, Vulkan,
Magma -- with the possibilty of a common decoder agnostic values
for ASG.

The benefit of the virtio-gpu capabilities is they are also cached
per VM instance, so probably a bit faster than the
rcGetHostExtensionsString method.

BUG=294229723
TEST=compile

Change-Id: I4b4bac4743293b5f7de2d06e426a8b1abb9b0d8f