gfxstream: mesa: import Mesa
Mesa has Linux WSI implementations that can be used for
gfxstream.  We have to first import the parts of Mesa that
we find useful.

Just modified

- guest/src/mesa/meson.build

to remove extra project args.

- guest/src/mesa/src/meson.build

to not automatically call subdir in all directories.

This also requires the VK_NO_NIR and Android build patches, which
are fairly upstreamable.  The strategy would be to merge all changes
required in the source code.

Generally, if you want to modify the Mesa subdir, you generally
will want to have a plan for upstreaming the change.
For example, here's the Mesa VK_NIR patches.

https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/26574

[not upstreamed yet; but do have a plan ;-)]

BUG=313466265
TEST=compile

Change-Id: Ib5a7546e3f8832ddba4de8e62afc80c669692444