gfxstream: nuke TcpStream
The last significant commit I can trace for this is
("b85b2756: migrate opengl and system from  development/tools")
from 2013.  Therefore, it's likely the TcpStream is dead code
and should be deleted.

Tcp streams are unlikely to work with VK, for example.

Note for the CMake-based host build,
`guest/OpenglSystemCommon/CMakeLists.txt" had to be hand-edited to
include the proper Vulkan include directories.  For some reason and
without this CL, typing "make" in goldfish-opengl omits this.  Since it
was a prior bug and the CMake-based host build should be deleted
anyways, just hand edited for now.

BUG=296903491
TEST=compile

Change-Id: I2254f1c2c0d5c38d0ec626d9d5d52fdac09d6a0a