gfxstream/guest: pass through Android process names
This allows complex process names such as "com.android.systemui"
and "android.hardware.graphics.composer3-service.ranchu" to be
be the debug name via the virtio-gpu interface.

This uses the CONTEXT_INIT ioctl to pass through the name.  Due
to TASK_COMM_LEN and the fact we initialize many times using
"RenderThread", the process name wasn't always propagated through
before.

This is useful for debugging and a host composition flow as well.

BUG=304997376
TEST=see process names host side

Change-Id: I2fc2fe1eaea1db45ab02dcb85e7890f1abed337b