Filter out swapchain maintenance feature if not supported on host
The Android framework's libvulkan implements this after ag/20821942.
However, the Android framework can't really mess with the create
info's pNext chain in the way Gfxstream can (since Gfxstream makes
a fully copy of known structs and filters the rest implicitly).
With this, the VkPhysicalDeviceSwapchainMaintenance1FeaturesEXT
struct is passed to the host regardless of whether or not the host
supports the feature because the libvulkan claims support.

This change updates the host to check if the feature was requested
and disables the feature if the host does not support it.

Bug: b/310231514
Test: dEQP-VK.api.external.memory.android_hardware_buffer.*
Change-Id: Id019e6af126964dc6332662d6f7833fe236d8ec1