Use old en/decoding for vkQueueSignalReleaseImage
We accidentially changed the xml definition for
vkQueueSignalReleaseImageANDROID. As a result it changed its encoding
and decoding, which breaks Android emulator when launching old system
images (API level 29~32).

To fix this, we overwrite it with the old definition so that it keeps
using the old encoding/decoding.

Note that API 33+ uses vkQueueSignalReleaseImageANDROIDAsyncGOOGLE which
does not have this problem.

Bug: 314333200
Test: API 29 on Mac m1 launch chrome
Change-Id: If3c5b9cb880ba7a415c98a172dd0fb43215f9b98