Remove workaround in when updating colorbuffer
This change removes a workaround in place when calling
updateColorBufferFromBytesLocked where the destination buffer was
explicitly set to have an original layout of TRANSFER_DST_OPTIMAL. The
intent of this workaround was to leverage undefined vulkan driver
behavior to avoid a risk that the destination buffer would be discarded
or overwritten inappropriately. That could happen because the underlying
(gralloc) buffer could be used by either vulkan or GL on the guest, leaving
the host's vulkan state unaware of the actual layout.

But since this function validates that the region provided aligns with
the destination buffer's size fairly early on and does not support
subrects, this data was bound to be overwritten anyway, making guarding
against discarded data unnecessary.

Thus the removal of this workaround fixes the vulkan validation layer
failures tied to the previously intended use of undefined behavior.

Bug: 323612547
Tested: Ran on emulator
Change-Id: I6ab327ab39a9009e978140587823ac16114f60e7