Bazel: Do not use cc_shared_library
It looks that the cc_shared_libary function will not link in the
objc_library dependencies. This in turn will result in missing symbols
when attempting to load the dylib inside of qemu.

We now use the "old" version of using cc_binary to create the dylib.
This requires us to slightly rewrite the rules to make sure all symbols
are present.

The bug contains additional google internal background information.

Test: Can launch qemu with gfxstream on mac os
Bug: 331235218
Change-Id: Ia634f9d6e1f9495d288a7c751ab176d33eb9247f