Allow static linking of libgfxstream_backend
The library can now be linked either statically or dynamically. This is
to minimize the storage usage. This library is quite large (~5MB), but
we can make it smaller by statically linking it and pruning unused
object files. With static linking this takes approximately 3MB.

Note that whether to link this library statically or dynamically is up
to the client.

Bug: 327141614
Test: m libgfxstream_backend
Change-Id: I19fbd2f544916296cb11312214b35e4b5fc62f69