Fix ColorBufferVk layout on snapshot load again
We have a situation that the color buffer vulkan images are transformed
into the wrong layout after snapshot load. Fix it by memorizing the
layout in snapshot.

We don't have a test for it yet as we don't really understand how to
trigger this behavior at unittest level.

Also recover layout for multi-sample images after snapshot loads. But we
still don't snapshot its content yet.

Bug: 332775761
Test: CanSubmitQueue
Change-Id: I8007f504d055041b11c3af3425cfbdbfac0553dd