gfxstream: always advertise device memory extension guest side
GfxstreamEnd2EndVkTest::DeviceMemoryReport always assumes the
memory report extension is present.  It's is filtered out when
sent host side, since for a virtual GPU this is quite difficult
to implement.

Mesa runtime is picky about physical device features.  So if
the test tries to enable device level extension without it
definitely existing, the test will fail.

The test can also be modified to check
VkPhysicalDeviceDeviceMemoryReportFeaturesEXT, but that's more
involved.  Work around this by always advertising the extension.

BUG=338270042
TEST=end2end tests pass

Change-Id: Ica3ba0f976bb0a1287121008537b2735f9276b70