Define MoltenVK functions for gfxstream
MoltenVK functions are not part of the official vulkan
and are being used directly from the shared library.
Auto code generation is not generating headers and
we need to manually add them in private defs.

This would enable some MoltenVK codepaths for gfxstream
and we'll replace it with VK_EXT_metal_objects later.

Bug: 332964912
Bug: 339094207
Test: build & run emulator on mac
Change-Id: Ia59cd270c342c38165d2f642d78391b8d07ce73f