gfxstream: guest: introduce GfxStreamConnectionManager
The HostConnection class pulls in:

  - GLES_v1 encoder
  - GLES_v2 encoder
  - renderControl encoder
  - vulkan encoder

Since HostConnection is required to talk the host, that
means the libgfxstream_vulkan.so has a GLES dependency.

This is unnecessary.  Ideally, libgfxstream_vulkan.so
only needs the Vulkan encoder dependency, and that is
the case for virtgpu.

For Goldfish, since it uses the "process pipe"
initialization logic is used, for Android builds
a renderControl encoder dependency is still required.

To increase the separation of these encoders, this
change adds a GfxStreamConnectionManager class.  It
has no explicit dependency on any API encoder, but
can store thread-local instances of them.

The HostConnection class may be implemented in terms
of the GfxStreamConnectionManager API.  For now, the
plan is just to convert libgfxstream_vulkan to use
the ConnectionManager.

BUG=311486792
TEST=CI + compile

Change-Id: I5562fd7b44b4d516c387b553081cd3ee88571ea0