Ensure vkAllocateMemory() does not have both import and export structs
After aosp/2973672, a `VkMemoryAllocateInfo` could end up with a
`VkImportMemoryFdInfoKHR`, `VkImportMemoryHostPointerInfoEXT`, and
`VkExportMemoryAllocateInfo` when running with SystemBlob enabled
(see http://gpaste/5463149521403904). Prior to aosp/2973672,
vkAllocateMemory depended on clearing out the always set `pNext`
field.

This change attempts to reorganize some of if/else if/else to ensure
that different branches do not need to clear out the pNext.

A couple minor diffs from clang-format as well.

Bug: b/364278658
Test: cvd start --gpu_mode=gfxstream
Test: cvd start --gpu_mode=gfxstream_guest_angle
Test: cvd start --gpu_mode=gfxstream_guest_angle \
                --gpu_vhost_user_mode=on
Test: cvd start --gpu_mode=gfxstream_guest_angle_host_swiftshader
Test: cvd start --gpu_mode=gfxstream_guest_angle_host_swiftshader \
                --gpu_vhost_user_mode=on
Change-Id: Id5a463cba05f16919b5d752c74eb3f045043ae1f