[Vulkan Snapshot]: handle multi-allocation better
Do not delete the ApiInfo that keeps the trace
to create multiple Vulkan Objects, if it still
has some objects left in its createdHandles.

Only do that after all the handles are removed,
so that the remaining handles will still be
loaded and usable; Otherwise, the emulator will
crash later on if one of the handles is used.

This will potentiall lead to resource leakage,
and it should be handled shortly.

Test: create an vulkan app, allocate command
buffers in one batch, e.g, 4, free the last
command buffer; use the remaining 3 for rendering
and save snapshot; it should not crash in save
and it should load fine.

Bug: 372242705

Change-Id: If671e95acae9cf49e490d9e46ad393cddf6a3e49