Add liblog to static libraries linked with perfetto
libperfetto_client_experimental is a static library that depends on the
liblog shared_library.

Shared libraries and binaries that depend on
libperfetto_client_experimental (directly or indirectly, **even through
static libraries**) need to include liblog in their shared_libs, because
soong doesn't actually propagate shared_libs dependecies (this appears
to be by design, see (sorry, google internal only)
https://yaqs.corp.google.com/eng/q/7654326063064940544).

Static libraries, on the other hand, only need to list liblog for the
headers that are exported by perfetto. There are two solutions for this:

1. `libperfetto_client_experimental` uses `export_shared_lib_headers`.
   This is the approach taken by https://r.android.com/3254919.
2. Every static_library that depends on
   `libperfetto_client_experimental` (directly or indirectly) lists
   `liblog` as `shared_libs`.

This does 2 instead of 1. I think it's better because:
* `libperfetto_client_experimental` is used to build prebuilts for ART.
  With `export_shared_lib_headers: liblog`, liblog headers would be
  copied twice in ART, as part of libperfetto_client_experimental and
  liblog. That would create conflicts.
* Since binaries and shared libraries that depend on
  libperfetto_client_experimental ultimately need liblog, I think it's
  better that the nodes in the middle of the dependency chain (the
  static libraries) depend on liblog as well. It's easier to reason
  about the dependency this way.

Bug: 374113246
Test: cd hardware/google/gfxstream && mma with https://r.android.com/3313296
Change-Id: Ie90bfdd5b022e5e96dc718f184d19a5977580b91