Remove snapshot tests
These tests have been relying on the fact that frontend did not
actually change the virtio objects used for transport (the virtio
gpu resources used for either the pipe stream or ASG stream) across
snapshot() and restore(). This will no longer be the case with upcoming
changes as the frontend will start to actually create resources during
restore().

The end2end testing framework is kind of inadequate to handle
snapshot testing as is because it can not remap the "host" resources
into existing "guest" objects. This works for full VM testing because
KVM can remap newly allocate resources (from restore()) back into an
existing guest address space.

The short term plan for testing is to start to check in simple VK
apps that can be used with Cuttlefish/Goldfish during integration
tests that have the ability to fully launch a VM.

It could be possible to revive snapshot testing in the end2end framework
using a custom page fault handler. The virtio pipe resource buffers and
the virtio blob resource shmem could be hidden behind another layer of
indirection in some kind of "resource indirection table" and a custom
page fault handler could potentially re-route reads/writes from/to the
memory that is shared with the "guest". Its not clear how much work this
would require or if it would fully work so let's just remove these
end2end tests for now and rely on integration tests with full virtual
devices.

Bug: b/369615058
Test: build
Change-Id: Id6eb1385315fad95ef89911b91fa0c835f769fb2