Initial implementation of multiple queue emulation
Add a virtual queue to queue families with graphics support to
advertise multiple queues to the guest when required, to run
SkiaVK on Android. This CL is not tracking semaphores and fences
on those queues to optimize the wait idle calls, and those
changes will be implemented later. Timeline semaphores are not
expected to be supported on virtual queues.

Virtual queue handle will have their last bit set to 1 to avoid
locking for mQueueInfo array during unboxing. A simple error logic
is implemented to handle possible invalid handles coming from the
GPU, but we can do a lightweight tracking to ensure wider support.

Use ANDROID_EMU_VK_ENABLE_VIRTUAL_QUEUE=1 to enable the virtual
queue on a device with only a single graphics queue, or when using
swiftshader. It'll be enabled by default once it's ready.

Bug: 327230380
Test: deqp *.multi_queue.*
Change-Id: Ib428ff63e9dc6dd7aa0d10f2daffac5aeb916280