Separate and rename snapshot "ApiInfo" structs/types
... to make these more available for use in VkDecoder
to help fix up some thread safety issues.

Renames the types a little bit to include "snapshot" in the names,
to make the distinction between info about the API and about a specific
call to the API, and to clarify that the "manager" is not a "trace" (
currently named `ApiTrace`) but is a collection of "api calls".

Bug: b/383139165
Test: m
Change-Id: I5ea8dbcc3a0590acfd74e6e66ee94b0ad5fb0535