vulkan: fix resource leak when destroy device
Commonly, the guest calls vkDestroyDevice before
calling vkDestroyInstance. However, if guest
forgets to destroy some device specific resources,
they cannot be cleaned up on the host side,as
the device is already destroyed, and the parent
instance has no record of it when doing final
sweep of garbage.

This cl does some refactoring to make sure
the device resouces are cleaned up when guest
calls vkDestroyDevice.

Bug: 370344466

Change-Id: Ieca0765942e791f91bf452d0553395e2c4a0b506