Update delayed delete callbacks to happen without global lock
The delayed delete callback is currently only used for
VkPipelineLayout-s. aosp/3449121 updated the delayed delete
callback to go through VkDecoderGlobalState instead of directly
to the host Vulkan driver. This works currently with the
following callstack:

  * VkDecoderGlobalState::on_vkDestroyPipelineLayout()
  * <std::function for VkPipelineLayout>
  * processDelayedRemovesGlobalStateLocked()
  * VkDecoderGlobalState::on_vkDestroyDevice()

because of the use of std::recursive_mutex. However,
std::recursive_mutex does not support thready safety annotations.
This change breaks the need for the std::recursive_mutex by making
the processing of the delayed deletes happen outside of the
VkDecoderGlobalState mutex lock so that subsequent changes can
start to update VkDecoberGlobalState's members with thread safety
annotations.

Bug: b/383139165
Test: cvd create --gpu_mode=gfxstream_guest_angle_host_swiftshader
Change-Id: I7e068b9b879547a7a0904d44da5213763541c437